requirements-full.txt
sentence-transformers>=2.7.0
nltk>=3.9.1
tiktoken>=0.8.0
numpy>=1.26.4
scipy>=1.14.1
scikit-learn>=1.5.2
pandas>=2.2.3
openai>=1.54.4
datasets>=3.1.0
ragas>=0.2.15
langchain>=0.3.7
langchain-openai>=0.2.8
langchain-community>=0.3.7
transformers>=4.46.3
torch>=2.5.1
faiss-cpu>=1.9.0
chromadb>=0.5.23
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.36
pypdf2>=3.0.1